Make ResponseObject.serialize() internal.
The ResponseObject.serialize() method depends on the ResponseObject
having been bound to an action descriptor, which it isn't if it is
instantiated by an action method.  Since serialize() is really only
intended to be called by Appathy itself, we make it internal by
changing its name to _serialize().

The binding comment above also applies to the code property, so this
will have to be revisited again at some point.  A potential solution
is to simply include the action descriptor in the request object
somewhere; then we would no longer need the kwargs in the
ResponseObject constructor, as it could derive the requisite
descriptor from the request, which has to be passed in anyway.