Allow bailing out with webob.Response
When an action method or extension returns a webob.Response, the
desired behavior is to bail out of the extension processing chain
and immediately return webob.Response.  We have been arranging for
this by simply having ActionDescriptor.wrap() raise the response
if it's a webob.Response object, but that only works if the
response is *also* a webob.exc.HTTPException, which descends from
Exception.

To allow for this bailing out behavior when the response is a
simple webob.Response object, we create a new AppathyResponse
special-case exception to ferry the response up to the Application
class dispatch() method.