Simplify extension processing
As for 'resource.*', it is not possible for there to be two different
'extend.*' configuration values affecting the same resource to be
passed into the function.  This allows considerable simplification of
the extension processing.

Also fixed a minor bug: in the final extension processing loop, we
try to import 'value', which is meaningless in this context; we
actually need to import 'ext_class'.