Add 'context' to the Request class
Create a custom Request class based on the webob Request class with a
single addition: a 'context' attribute mapped to the 'appathy.context'
environment key.  This key could be set by authentication middleware,
or by request preprocessors, and would then be accessible to
controllers very conveniently, as an attribute of the request object.